pyproject.toml
[project]
name = "test-mem-prof"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.114.1",
    "matplotlib>=3.9.2",
    "memory-profiler>=0.61.0",
    "tensorflow-cpu>=2.17.0",
    "uvicorn>=0.30.6",
]
